bugfix imagick adapter
git-svn-id: http://www.pimcore.org/svn/pimcore/private/core/trunk@4004 1f8fe7d8-47f0-464c-8d0a-336f4953ab05

+++ pimcore/lib/Pimcore/Image/Adapter/Imagick.php
@@ -110,8 +110,10 @@ class Pimcore_Image_Adapter_Imagick extends Pimcore_Image_Adapter {
      * @return  void
      */
     protected function destroy() {
-        $this->resource->clear();
-        $this->resource->destroy();
+        if($this->resource) {
+            $this->resource->clear();
+            $this->resource->destroy();
+        }
     }
 
     /**
